fix(ci+eudr): clippy/fmt clean + bound EUDR visual-evidence latency
CI fixes (CI runs cargo fmt --check + clippy -D warnings; both were failing
on the new EUDR/SAR/OPERA code):
- cargo fmt --all across emem-api-rest, physics, stac, opera_dist.
- clippy: unnecessary i64->i64 casts (opera_dist), too_many_arguments allow
  on stac::search_many_at, sort_by->sort_by_key, `>= x+1`->`> x`, negated
  partial-cmp on quantity_kg (-> is_finite + <= 0.0), doc-list indentation in
  build_plot_forest_context.

EUDR visual-evidence latency (the live 504 root cause beyond the router fix):
the per-year timeline materialized NDVI+S1 across ALL sample_cells (up to 256)
SEQUENTIALLY per band per year — 256 × 7y × 2 bands of stacked awaits blew even
the 240s EUDR budget on a cold plot. And forest_context looped 64 cells
sequentially. Both now:
- cap the cells to a representative sample (EMEM_EUDR_VISUAL_CELLS default 8,
  EMEM_EUDR_CONTEXT_CELLS default 16) — medians/distributions don't need every
  cell; scene PNGs still use the first url_cell_cap of them;
- warm the first cell (cold STAC+COG), then fan the rest out CONCURRENTLY so
  the per-tile TILE_CACHE/PROFILE_CACHE single-flight overlaps the warm reads
  instead of stacking sequential awaits.
Every fetch stays per-band-timeout-wrapped + under the global materialize
semaphore + the dedicated EUDR route budget. Bounded, no unbounded fan-out.

Full workspace suite green; clippy + fmt clean.

+++ crates/emem-fetch/src/opera_dist.rs
@@ -164,8 +164,8 @@ fn add_days_to_epoch(epoch: (i32, u32, u32), days: i64) -> (i32, u32, u32) {
 fn days_from_civil(y: i32, m: u32, d: u32) -> i64 {
     let y = if m <= 2 { y - 1 } else { y } as i64;
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64;
-    let mp = ((m as i64 + 9) % 12) as i64;
+    let yoe = y - era * 400;
+    let mp = (m as i64 + 9) % 12;
     let doy = (153 * mp + 2) / 5 + d as i64 - 1;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
     era * 146097 + doe - 719468
